requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
mypy-extensions==0.4.3
pathspec==0.8.0
regex==2020.10.28
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Flask==1.1.2
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
marshmallow-sqlalchemy==0.22.3
